+++ src/video/android/SDL_androidmouse.c
@@ -121,7 +121,7 @@ static SDL_Cursor *Android_CreateEmptyCursor()
     if (empty_cursor == NULL) {
         SDL_Surface *empty_surface = SDL_CreateRGBSurfaceWithFormat(0, 1, 1, 32, SDL_PIXELFORMAT_ARGB8888);
         if (empty_surface) {
-            SDL_memset(empty_surface->pixels, 0, empty_surface->h * empty_surface->pitch);
+            SDL_memset(empty_surface->pixels, 0, (size_t)empty_surface->h * empty_surface->pitch);
             empty_cursor = Android_CreateCursor(empty_surface, 0, 0);
             SDL_FreeSurface(empty_surface);
         }
